feat(spi): enhance security and persistence abstractions with improved API clarity and credential safety

+++ exeris-kernel-spi/src/main/java/eu/exeris/kernel/spi/persistence/PersistenceConnection.java
@@ -104,7 +104,7 @@ public interface PersistenceConnection extends AutoCloseable {
      * @return number of rows affected
      * @throws PersistenceProviderException on execution failure
      */
-    int executeUpdate(String sql);
+    long executeUpdate(String sql);
 
     // =========================================================================
     // Transaction Control
